parte de ubicacion, a donde ir, seleccion de pines, y casi toda la parte de avanzar terminada. Faltaria configurar el PWM y seleccionar las salidas para las funciones avanzar, apagar_derecha y apagar_izquierda. Faltaria el codigo girar y faltaria toda laparte de decision (para desp)

+++ Giro.c
@@ -11,7 +11,7 @@
 #define GIRO_180   3
 
 // Devuelve la dirección hacia donde hay que ir según la diferencia entre casillas
-int obtener_orientacion_N(int casillaActual, int casilla_N) {
+int8_t obtener_orientacion_N(int8_t casillaActual, int8_t casilla_N) {
     if (casilla_N == casillaActual + 1) return OESTE;
     if (casilla_N == casillaActual - 1) return ESTE;
     if (casilla_N == casillaActual + 4) return NORTE;
@@ -21,7 +21,7 @@ int obtener_orientacion_N(int casillaActual, int casilla_N) {
 }
 
 // Calcula el giro que debe hacer el autito para pasar de su orientación actual a la deseada
-int obtenerGiro(int orientacionActual, int orientacion_N) {
+int8_t obtenerGiro(int8_t orientacionActual, int8_t orientacion_N) {
     int diferencia = (orientacion_N - orientacionActual + 4) % 4; //el %4 se queda con el resto de la divsion por 4
 
     switch (diferencia) {
